Add "dev:proxy" script to the client's package.json
This allows us to test the full app with the express server running on port 3001 and the client using a proxy to forward all relative apis to such port.
This is in contrast with running the server in dev mode and having it serve the built React app. We can thus test the React app with out building it before, with the server running.